refactor: improve file struct
